refactor: split TsdRadarControl and MapDataService into partial classes
TsdRadarControl.cs (1,862 lines) → 5 files:
- Properties: StyledProperty declarations and CLR wrappers
- Projection: Lambert Conformal Conic math
- Input: pointer handling, context menus, keyboard shortcuts
- Rendering: Render() and all Draw* methods
- Core: fields, events, lifecycle, geometry/cache management

MapDataService.cs (1,059 lines) → 4 files:
- Core: nav data loading, lookups, resource helpers
- Boundaries: state/country/TRACON/ARTCC loading, GeoJSON parsing
- Routes: airway loading, route resolution, altitude estimation
- Sectors: ARTCC sector loading and spatial queries

No interfaces changed. No behavioral changes. Purely file organization
using C# partial classes.
