+++ Views/TsdRadarControl.cs
@@ -1,10 +1,8 @@
 ﻿using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Input;
 using Avalonia.Media;
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.Linq;
 using System.Threading;
@@ -12,16 +10,19 @@ using vTFMS.Models;
 
 namespace vTFMS.Views;
 
-public class TsdRadarControl : Control
+/// <summary>
+/// Custom radar display control that renders aviation map overlays,
+/// weather radar, VATSIM flight data, and interactive map items.
+///
+/// Split across partial class files for maintainability:
+///   TsdRadarControl.cs             — Core state, events, geometry/cache management
+///   TsdRadarControl.Properties.cs  — Avalonia StyledProperties and CLR wrappers
+///   TsdRadarControl.Projection.cs  — Lambert Conformal Conic projection math
+///   TsdRadarControl.Input.cs       — Pointer, context menu, and keyboard handling
+///   TsdRadarControl.Render.cs      — Render() and all Draw* methods
+/// </summary>
+public partial class TsdRadarControl : Control
 {
-    // =========================================================================
-    // Constants
-    // =========================================================================
-
-    // Lambert Conformal Conic standard parallels for US
-    private const double Phi1 = 33.0 * Math.PI / 180.0;
-    private const double Phi2 = 45.0 * Math.PI / 180.0;
-
     // =========================================================================
     // Cached geometry state
     // =========================================================================
@@ -103,372 +104,6 @@ public class TsdRadarControl : Control
     public event EventHandler<string>? GenericMenuCommandRequested;
 
     // =========================================================================
-    // Styled Properties
-    // =========================================================================
-
-    #region Styled Properties
-
-    public static readonly StyledProperty<ObservableCollection<StateBoundary>>
-        StateBoundariesProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<StateBoundary>>(
-            nameof(StateBoundaries),
-            new ObservableCollection<StateBoundary>());
-
-    public static readonly StyledProperty<ObservableCollection<StateBoundary>>
-        CountryBoundariesProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<StateBoundary>>(
-            nameof(CountryBoundaries),
-            new ObservableCollection<StateBoundary>());
-
-    public static readonly StyledProperty<ObservableCollection<SectorBoundary>>
-        SectorsProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<SectorBoundary>>(
-            nameof(Sectors),
-            new ObservableCollection<SectorBoundary>());
-
-    public static readonly StyledProperty<ObservableCollection<Airport>>
-        AirportsProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<Airport>>(
-            nameof(Airports),
-            new ObservableCollection<Airport>());
-
-    public static readonly StyledProperty<ObservableCollection<Navaid>>
-        NavaidsProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<Navaid>>(
-            nameof(Navaids),
-            new ObservableCollection<Navaid>());
-
-    public static readonly StyledProperty<ObservableCollection<Waypoint>>
-        WaypointsProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<Waypoint>>(
-            nameof(Waypoints),
-            new ObservableCollection<Waypoint>());
-
-    public static readonly StyledProperty<ObservableCollection<MapItem>>
-        ActiveMapItemsProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<MapItem>>(
-            nameof(ActiveMapItems),
-            new ObservableCollection<MapItem>());
-
-    public static readonly StyledProperty<ObservableCollection<VatsimPilot>>
-        VisiblePilotsProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<VatsimPilot>>(
-            nameof(VisiblePilots),
-            new ObservableCollection<VatsimPilot>());
-
-    public static readonly StyledProperty<ObservableCollection<ArtccBoundary>>
-        ArtccBoundariesProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<ArtccBoundary>>(
-            nameof(ArtccBoundaries),
-            new ObservableCollection<ArtccBoundary>());
-
-    public static readonly StyledProperty<bool> ShowStateBoundariesProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowStateBoundaries), true);
-
-    public static readonly StyledProperty<bool> ShowCountryBoundariesProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowCountryBoundaries), true);
-
-    public static readonly StyledProperty<bool> ShowArtccProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowArtcc), false);
-
-    public static readonly StyledProperty<bool> ShowWeatherProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowWeather), false);
-
-    public static readonly StyledProperty<byte[]?> RadarImageDataProperty =
-        AvaloniaProperty.Register<TsdRadarControl, byte[]?>(
-            nameof(RadarImageData));
-
-    public static readonly StyledProperty<double> RadarMinLatProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(nameof(RadarMinLat));
-
-    public static readonly StyledProperty<double> RadarMinLonProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(nameof(RadarMinLon));
-
-    public static readonly StyledProperty<double> RadarMaxLatProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(nameof(RadarMaxLat));
-
-    public static readonly StyledProperty<double> RadarMaxLonProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(nameof(RadarMaxLon));
-
-    public static readonly StyledProperty<double> RadarOpacityProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(
-            nameof(RadarOpacity), 0.7);
-
-    public static readonly StyledProperty<double> CenterLatProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(
-            nameof(CenterLat), 39.5);
-
-    public static readonly StyledProperty<double> CenterLonProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(
-            nameof(CenterLon), -98.35);
-
-    public static readonly StyledProperty<double> ZoomLevelProperty =
-        AvaloniaProperty.Register<TsdRadarControl, double>(
-            nameof(ZoomLevel), 1.0);
-
-    public static readonly StyledProperty<bool> ShowAirportsProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowAirports), true);
-
-    public static readonly StyledProperty<bool> ShowWaypointsProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowWaypoints), false);
-
-    public static readonly StyledProperty<bool> ShowVorsProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowVors), true);
-
-    public static readonly StyledProperty<bool> ShowNdbsProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowNdbs), true);
-
-    public static readonly StyledProperty<DisplaySettings>
-        DisplaySettingsProperty =
-        AvaloniaProperty.Register<TsdRadarControl, DisplaySettings>(
-            nameof(DisplaySettings), new DisplaySettings());
-
-    public static readonly StyledProperty<ObservableCollection<RangeRingConfig>>
-        RangeRingsProperty =
-        AvaloniaProperty.Register<TsdRadarControl,
-            ObservableCollection<RangeRingConfig>>(
-            nameof(RangeRings),
-            new ObservableCollection<RangeRingConfig>());
-
-    public static readonly StyledProperty<bool> ShowAllAircraftProperty =
-        AvaloniaProperty.Register<TsdRadarControl, bool>(
-            nameof(ShowAllAircraft), false);
-
-    #endregion
-
-    // =========================================================================
-    // CLR Property Wrappers
-    // =========================================================================
-
-    #region CLR Property Wrappers
-
-    public ObservableCollection<StateBoundary> StateBoundaries
-    {
-        get => GetValue(StateBoundariesProperty);
-        set => SetValue(StateBoundariesProperty, value);
-    }
-
-    public ObservableCollection<StateBoundary> CountryBoundaries
-    {
-        get => GetValue(CountryBoundariesProperty);
-        set => SetValue(CountryBoundariesProperty, value);
-    }
-
-    public ObservableCollection<SectorBoundary> Sectors
-    {
-        get => GetValue(SectorsProperty);
-        set => SetValue(SectorsProperty, value);
-    }
-
-    public ObservableCollection<Airport> Airports
-    {
-        get => GetValue(AirportsProperty);
-        set => SetValue(AirportsProperty, value);
-    }
-
-    public ObservableCollection<Navaid> Navaids
-    {
-        get => GetValue(NavaidsProperty);
-        set => SetValue(NavaidsProperty, value);
-    }
-
-    public ObservableCollection<Waypoint> Waypoints
-    {
-        get => GetValue(WaypointsProperty);
-        set => SetValue(WaypointsProperty, value);
-    }
-
-    public ObservableCollection<MapItem> ActiveMapItems
-    {
-        get => GetValue(ActiveMapItemsProperty);
-        set => SetValue(ActiveMapItemsProperty, value);
-    }
-
-    public ObservableCollection<VatsimPilot> VisiblePilots
-    {
-        get => GetValue(VisiblePilotsProperty);
-        set => SetValue(VisiblePilotsProperty, value);
-    }
-
-    public ObservableCollection<ArtccBoundary> ArtccBoundaries
-    {
-        get => GetValue(ArtccBoundariesProperty);
-        set => SetValue(ArtccBoundariesProperty, value);
-    }
-
-    public bool ShowStateBoundaries
-    {
-        get => GetValue(ShowStateBoundariesProperty);
-        set => SetValue(ShowStateBoundariesProperty, value);
-    }
-
-    public bool ShowCountryBoundaries
-    {
-        get => GetValue(ShowCountryBoundariesProperty);
-        set => SetValue(ShowCountryBoundariesProperty, value);
-    }
-
-    public bool ShowArtcc
-    {
-        get => GetValue(ShowArtccProperty);
-        set => SetValue(ShowArtccProperty, value);
-    }
-
-    public bool ShowWeather
-    {
-        get => GetValue(ShowWeatherProperty);
-        set => SetValue(ShowWeatherProperty, value);
-    }
-
-    public byte[]? RadarImageData
-    {
-        get => GetValue(RadarImageDataProperty);
-        set => SetValue(RadarImageDataProperty, value);
-    }
-
-    public double RadarMinLat
-    {
-        get => GetValue(RadarMinLatProperty);
-        set => SetValue(RadarMinLatProperty, value);
-    }
-
-    public double RadarMinLon
-    {
-        get => GetValue(RadarMinLonProperty);
-        set => SetValue(RadarMinLonProperty, value);
-    }
-
-    public double RadarMaxLat
-    {
-        get => GetValue(RadarMaxLatProperty);
-        set => SetValue(RadarMaxLatProperty, value);
-    }
-
-    public double RadarMaxLon
-    {
-        get => GetValue(RadarMaxLonProperty);
-        set => SetValue(RadarMaxLonProperty, value);
-    }
-
-    public double RadarOpacity
-    {
-        get => GetValue(RadarOpacityProperty);
-        set => SetValue(RadarOpacityProperty, value);
-    }
-
-    public double CenterLat
-    {
-        get => GetValue(CenterLatProperty);
-        set => SetValue(CenterLatProperty, value);
-    }
-
-    public double CenterLon
-    {
-        get => GetValue(CenterLonProperty);
-        set => SetValue(CenterLonProperty, value);
-    }
-
-    public double ZoomLevel
-    {
-        get => GetValue(ZoomLevelProperty);
-        set => SetValue(ZoomLevelProperty, value);
-    }
-
-    public bool ShowAirports
-    {
-        get => GetValue(ShowAirportsProperty);
-        set => SetValue(ShowAirportsProperty, value);
-    }
-
-    public bool ShowWaypoints
-    {
-        get => GetValue(ShowWaypointsProperty);
-        set => SetValue(ShowWaypointsProperty, value);
-    }
-
-    public bool ShowVors
-    {
-        get => GetValue(ShowVorsProperty);
-        set => SetValue(ShowVorsProperty, value);
-    }
-
-    public bool ShowNdbs
-    {
-        get => GetValue(ShowNdbsProperty);
-        set => SetValue(ShowNdbsProperty, value);
-    }
-
-    public DisplaySettings DisplaySettings
-    {
-        get => GetValue(DisplaySettingsProperty);
-        set => SetValue(DisplaySettingsProperty, value);
-    }
-
-    public ObservableCollection<RangeRingConfig> RangeRings
-    {
-        get => GetValue(RangeRingsProperty);
-        set => SetValue(RangeRingsProperty, value);
-    }
-
-    public bool ShowAllAircraft
-    {
-        get => GetValue(ShowAllAircraftProperty);
-        set => SetValue(ShowAllAircraftProperty, value);
-    }
-
-    #endregion
-
-    // =========================================================================
-    // Static constructor — render invalidation
-    // =========================================================================
-
-    static TsdRadarControl()
-    {
-        AffectsRender<TsdRadarControl>(
-            StateBoundariesProperty,
-            CountryBoundariesProperty,
-            SectorsProperty,
-            AirportsProperty,
-            NavaidsProperty,
-            WaypointsProperty,
-            CenterLatProperty,
-            CenterLonProperty,
-            ZoomLevelProperty,
-            ShowAirportsProperty,
-            ShowWaypointsProperty,
-            ShowVorsProperty,
-            ShowNdbsProperty,
-            ActiveMapItemsProperty,
-            VisiblePilotsProperty,
-            ShowStateBoundariesProperty,
-            ShowCountryBoundariesProperty,
-            ShowArtccProperty,
-            ArtccBoundariesProperty,
-            RadarImageDataProperty,
-            ShowWeatherProperty,
-            DisplaySettingsProperty,
-            RadarOpacityProperty,
-            RangeRingsProperty);
-    }
-
-    // =========================================================================
     // Constructor
     // =========================================================================
 
@@ -476,84 +111,6 @@ public class TsdRadarControl : Control
     {
         Focusable = true;
         RebuildRenderCache();
-    }
-
-    // =========================================================================
-    // Projection — Lambert Conformal Conic
-    // =========================================================================
-
-    private static (double x, double y) LccProject(
-        double lat, double lon,
-        double centerLat, double centerLon)
-    {
-        double phi = lat * Math.PI / 180.0;
-        double lam = lon * Math.PI / 180.0;
-        double phi0 = centerLat * Math.PI / 180.0;
-        double lam0 = centerLon * Math.PI / 180.0;
-
-        double n = Math.Log(Math.Cos(Phi1) / Math.Cos(Phi2)) /
-                   Math.Log(Math.Tan(Math.PI / 4.0 + Phi2 / 2.0) /
-                            Math.Tan(Math.PI / 4.0 + Phi1 / 2.0));
-
-        double F = Math.Cos(Phi1) *
-                   Math.Pow(Math.Tan(Math.PI / 4.0 + Phi1 / 2.0), n) / n;
-
-        double rho = F / Math.Pow(Math.Tan(Math.PI / 4.0 + phi / 2.0), n);
-        double rho0 = F / Math.Pow(Math.Tan(Math.PI / 4.0 + phi0 / 2.0), n);
-
-        double theta = n * (lam - lam0);
-
-        double x = rho * Math.Sin(theta);
-        double y = rho0 - rho * Math.Cos(theta);
-
-        return (x, y);
-    }
-
-    private static (double lat, double lon) LccInverse(
-        double px, double py,
-        double centerLat, double centerLon)
-    {
-        double phi0 = centerLat * Math.PI / 180.0;
-        double lam0 = centerLon * Math.PI / 180.0;
-
-        double n = Math.Log(Math.Cos(Phi1) / Math.Cos(Phi2)) /
-                   Math.Log(Math.Tan(Math.PI / 4.0 + Phi2 / 2.0) /
-                            Math.Tan(Math.PI / 4.0 + Phi1 / 2.0));
-
-        double F = Math.Cos(Phi1) *
-                   Math.Pow(Math.Tan(Math.PI / 4.0 + Phi1 / 2.0), n) / n;
-
-        double rho0 = F / Math.Pow(Math.Tan(Math.PI / 4.0 + phi0 / 2.0), n);
-
-        double rho = Math.Sign(n) * Math.Sqrt(
-            px * px + (rho0 - py) * (rho0 - py));
-        double theta = Math.Atan2(px, rho0 - py);
-
-        double phi = 2.0 * Math.Atan(Math.Pow(F / rho, 1.0 / n))
-                     - Math.PI / 2.0;
-        double lam = lam0 + theta / n;
-
-        return (phi * 180.0 / Math.PI, lam * 180.0 / Math.PI);
-    }
-
-    private Point LatLonToScreen(double lat, double lon,
-                                  double width, double height)
-    {
-        double scale = Math.Min(width, height) * 0.45 * ZoomLevel;
-
-        var (px, py) = LccProject(lat, lon, CenterLat, CenterLon);
-        var (cx, cy) = LccProject(CenterLat, CenterLon, CenterLat, CenterLon);
-
-        double x = width / 2.0 + (px - cx) * scale / (Math.PI / 180.0) / 57.0;
-        double y = height / 2.0 - (py - cy) * scale / (Math.PI / 180.0) / 57.0;
-
-        return new Point(x, y);
-    }
-
-    private bool IsOnScreen(Point pt, double width, double height)
-    {
-        return pt.X >= -20 && pt.X <= width + 20 &&
-               pt.Y >= -20 && pt.Y <= height + 20;
     }
 
     // =========================================================================
@@ -638,426 +195,6 @@ public class TsdRadarControl : Control
     {
         _geometriesDirty = true;
         InvalidateVisual();
-    }
-
-    // =========================================================================
-    // Input — Pointer (hover, right-click context menus)
-    // =========================================================================
-
-    protected override void OnPointerMoved(PointerEventArgs e)
-    {
-        base.OnPointerMoved(e);
-        _currentMousePosition = e.GetPosition(this);
-
-        var prevCallsign = _hoveredCallsign;
-        _hoveredCallsign = null;
-        _hoveredPilot = null;
-
-        foreach (var pilot in VisiblePilots)
-        {
-            var pt = LatLonToScreen(pilot.Lat, pilot.Lon,
-                Bounds.Width, Bounds.Height);
-            double dx = _currentMousePosition.X - pt.X;
-            double dy = _currentMousePosition.Y - pt.Y;
-
-            if (Math.Sqrt(dx * dx + dy * dy) <= 8.0)
-            {
-                _hoveredPilot = pilot;
-                _hoveredCallsign = pilot.Callsign;
-                break;
-            }
-        }
-
-        if (_hoveredCallsign != prevCallsign)
-            InvalidateVisual();
-    }
-
-    protected override void OnPointerEntered(PointerEventArgs e)
-    {
-        base.OnPointerEntered(e);
-        Focus();
-    }
-
-    protected override void OnPointerPressed(PointerPressedEventArgs e)
-    {
-        base.OnPointerPressed(e);
-
-        var point = e.GetCurrentPoint(this);
-        if (!point.Properties.IsRightButtonPressed) return;
-
-        var clickPos = point.Position;
-        var width = Bounds.Width;
-        var height = Bounds.Height;
-
-        // ── Hit-test 1: Flight icons ─────────────────────────
-        foreach (var pilot in VisiblePilots)
-        {
-            if (pilot.IsHidden) continue;
-            var pt = LatLonToScreen(pilot.Lat, pilot.Lon, width, height);
-            double dx = clickPos.X - pt.X;
-            double dy = clickPos.Y - pt.Y;
-
-            if (Math.Sqrt(dx * dx + dy * dy) <= 10.0)
-            {
-                ShowFlightContextMenu(pilot);
-                e.Handled = true;
-                return;
-            }
-        }
-
-        // ── Hit-test 2: Active map items ─────────────────────
-        foreach (var item in ActiveMapItems)
-        {
-            var pt = LatLonToScreen(item.Lat, item.Lon, width, height);
-            double dx = clickPos.X - pt.X;
-            double dy = clickPos.Y - pt.Y;
-
-            if (Math.Sqrt(dx * dx + dy * dy) <= 12.0)
-            {
-                ShowMapItemContextMenu(item);
-                e.Handled = true;
-                return;
-            }
-        }
-
-        // ── Hit-test 3: Empty space (generic) ────────────────
-        ShowGenericContextMenu(clickPos);
-        e.Handled = true;
-    }
-
-    // =========================================================================
-    // Context menus
-    // =========================================================================
-
-    /// <summary>Context menu for right-clicking a flight icon.</summary>
-    private void ShowFlightContextMenu(VatsimPilot pilot)
-    {
-        var menu = new ContextMenu();
-
-        var dataBlockItem = new MenuItem
-        {
-            Header = pilot.ShowDataBlock
-                ? "Hide Data Block" : "Show Data Block"
-        };
-        dataBlockItem.Click += (_, _) =>
-        {
-            pilot.ShowDataBlock = !pilot.ShowDataBlock;
-            InvalidateVisual();
-        };
-        menu.Items.Add(dataBlockItem);
-
-        var orgDestItem = new MenuItem
-        {
-            Header = pilot.ShowOrgDest
-                ? "Hide Org/Dest" : "Show Org/Dest"
-        };
-        orgDestItem.Click += (_, _) =>
-        {
-            pilot.ShowOrgDest = !pilot.ShowOrgDest;
-            InvalidateVisual();
-        };
-        menu.Items.Add(orgDestItem);
-
-        var drawRouteItem = new MenuItem
-        {
-            Header = (pilot.ManualDrawRoute || pilot.MatchedDrawRoute)
-                ? "Hide Route" : "Draw Route"
-        };
-        drawRouteItem.Click += (_, _) =>
-        {
-            pilot.ManualDrawRoute = !pilot.ManualDrawRoute;
-            if (pilot.ManualDrawRoute &&
-                pilot.ParsedRoute.Count == 0 &&
-                !string.IsNullOrWhiteSpace(pilot.Route))
-            {
-                RouteResolveRequested?.Invoke(this, pilot);
-            }
-            InvalidateVisual();
-        };
-        menu.Items.Add(drawRouteItem);
-
-        var showRouteItem = new MenuItem
-        {
-            Header = (pilot.ManualShowRoute || pilot.MatchedShowRoute)
-                ? "Hide Route Text" : "Show Route Text"
-        };
-        showRouteItem.Click += (_, _) =>
-        {
-            pilot.ManualShowRoute = !pilot.ManualShowRoute;
-            InvalidateVisual();
-        };
-        menu.Items.Add(showRouteItem);
-
-        // Color submenu
-        var colorMenu = new MenuItem { Header = "Change Color" };
-        var colors = new Dictionary<string, string>
-        {
-            { "White",   "#FFFFFF" },
-            { "Cyan",    "#00FFFF" },
-            { "Green",   "#00FF00" },
-            { "Yellow",  "#FFFF00" },
-            { "Orange",  "#FFA500" },
-            { "Red",     "#FF0000" },
-            { "Magenta", "#FF00FF" },
-            { "Reset",   "" }
-        };
-
-        foreach (var (name, hex) in colors)
-        {
-            var colorItem = new MenuItem { Header = name };
-            var capturedHex = hex;
-            colorItem.Click += (_, _) =>
-            {
-                pilot.ColorOverride =
-                    string.IsNullOrEmpty(capturedHex) ? null : capturedHex;
-                InvalidateVisual();
-            };
-            colorMenu.Items.Add(colorItem);
-        }
-        menu.Items.Add(colorMenu);
-
-        menu.Items.Add(new Separator());
-
-        var deleteItem = new MenuItem { Header = "Delete Icon" };
-        deleteItem.Click += (_, _) =>
-        {
-            pilot.IsHidden = true;
-            InvalidateVisual();
-        };
-        menu.Items.Add(deleteItem);
-
-        this.ContextMenu = menu;
-        menu.Open(this);
-    }
-
-    /// <summary>Context menu for right-clicking an active map item.</summary>
-    private void ShowMapItemContextMenu(MapItem item)
-    {
-        var menu = new ContextMenu();
-
-        var labelItem = new MenuItem
-        {
-            Header = item.ShowLabel ? "Hide Label" : "Show Label"
-        };
-        labelItem.Click += (_, _) =>
-        {
-            item.ShowLabel = !item.ShowLabel;
-            InvalidateVisual();
-        };
-        menu.Items.Add(labelItem);
-
-        var centerItem = new MenuItem { Header = "Center On" };
-        centerItem.Click += (_, _) =>
-        {
-            _prevCenterLat = CenterLat;
-            _prevCenterLon = CenterLon;
-            _prevZoomLevel = ZoomLevel;
-
-            CenterLat = item.Lat;
-            CenterLon = item.Lon;
-            ScheduleRadarRefresh();
-        };
-        menu.Items.Add(centerItem);
-
-        menu.Items.Add(new Separator());
-
-        var removeItem = new MenuItem { Header = "Remove Item" };
-        removeItem.Click += (_, _) =>
-        {
-            MapItemRemoveRequested?.Invoke(this, item);
-        };
-        menu.Items.Add(removeItem);
-
-        this.ContextMenu = menu;
-        menu.Open(this);
-    }
-
-    /// <summary>Context menu for right-clicking empty radar space.</summary>
-    private void ShowGenericContextMenu(Point clickPos)
-    {
-        var menu = new ContextMenu();
-
-        var centerItem = new MenuItem { Header = "Center Here" };
-        centerItem.Click += (_, _) =>
-        {
-            _prevCenterLat = CenterLat;
-            _prevCenterLon = CenterLon;
-            _prevZoomLevel = ZoomLevel;
-
-            var width = Bounds.Width;
-            var height = Bounds.Height;
-            double scale = Math.Min(width, height) * 0.45 * ZoomLevel;
-
-            double dxProj = (clickPos.X - width / 2.0)
-                * (Math.PI / 180.0) * 57.0 / scale;
-            double dyProj = (clickPos.Y - height / 2.0)
-                * (Math.PI / 180.0) * 57.0 / scale;
-
-            var (newLat, newLon) = LccInverse(
-                dxProj, -dyProj, CenterLat, CenterLon);
-
-            CenterLat = newLat;
-            CenterLon = newLon;
-            ScheduleRadarRefresh();
-        };
-        menu.Items.Add(centerItem);
-
-        var undoItem = new MenuItem { Header = "Undo Move/Zoom" };
-        undoItem.Click += (_, _) =>
-        {
-            var tempLat = CenterLat;
-            var tempLon = CenterLon;
-            var tempZoom = ZoomLevel;
-
-            CenterLat = _prevCenterLat;
-            CenterLon = _prevCenterLon;
-            ZoomLevel = _prevZoomLevel;
-
-            _prevCenterLat = tempLat;
-            _prevCenterLon = tempLon;
-            _prevZoomLevel = tempZoom;
-
-            ScheduleRadarRefresh();
-        };
-        menu.Items.Add(undoItem);
-
-        menu.Items.Add(new Separator());
-
-        var selectFlightsItem = new MenuItem { Header = "Select Flights..." };
-        selectFlightsItem.Click += (_, _) =>
-            GenericMenuCommandRequested?.Invoke(this, "SelectFlights");
-        menu.Items.Add(selectFlightsItem);
-
-        var showMapItem = new MenuItem { Header = "Show Map Item..." };
-        showMapItem.Click += (_, _) =>
-            GenericMenuCommandRequested?.Invoke(this, "ShowMapItem");
-        menu.Items.Add(showMapItem);
-
-        var rangeRingsItem = new MenuItem { Header = "Range Rings..." };
-        rangeRingsItem.Click += (_, _) =>
-            GenericMenuCommandRequested?.Invoke(this, "RangeRings");
-        menu.Items.Add(rangeRingsItem);
-
-        this.ContextMenu = menu;
-        menu.Open(this);
-    }
-
-    // =========================================================================
-    // Input — Keyboard (quick keys)
-    // =========================================================================
-
-    protected override void OnKeyDown(KeyEventArgs e)
-    {
-        base.OnKeyDown(e);
-
-        var width = Bounds.Width;
-        var height = Bounds.Height;
-
-        switch (e.Key)
-        {
-            case Key.A:
-                ShowArtcc = !ShowArtcc;
-                e.Handled = true;
-                break;
-
-            case Key.B:
-                bool boundariesOn = ShowStateBoundaries || ShowCountryBoundaries;
-                ShowStateBoundaries = !boundariesOn;
-                ShowCountryBoundaries = !boundariesOn;
-                e.Handled = true;
-                break;
-
-            case Key.D:
-                _geometriesDirty = true;
-                InvalidateVisual();
-                e.Handled = true;
-                break;
-
-            case Key.F:
-                ShowAllAircraft = !ShowAllAircraft;
-                e.Handled = true;
-                break;
-
-            case Key.I:
-                _prevCenterLat = CenterLat;
-                _prevCenterLon = CenterLon;
-                _prevZoomLevel = ZoomLevel;
-
-                CenterLat = 38.5;
-                CenterLon = -96.0;
-                ZoomLevel = 4.0;
-                ShowStateBoundaries = true;
-                ShowCountryBoundaries = true;
-
-                e.Handled = true;
-                ScheduleRadarRefresh();
-                break;
-
-            case Key.M:
-                _prevCenterLat = CenterLat;
-                _prevCenterLon = CenterLon;
-                _prevZoomLevel = ZoomLevel;
-
-                double scale = Math.Min(width, height) * 0.45 * ZoomLevel;
-
-                double dxProj = (_currentMousePosition.X - width / 2.0)
-                    * (Math.PI / 180.0) * 57.0 / scale;
-                double dyProj = (_currentMousePosition.Y - height / 2.0)
-                    * (Math.PI / 180.0) * 57.0 / scale;
-
-                var (newLat, newLon) = LccInverse(
-                    dxProj, -dyProj, CenterLat, CenterLon);
-
-                CenterLat = newLat;
-                CenterLon = newLon;
-                e.Handled = true;
-                ScheduleRadarRefresh();
-                break;
-
-            case Key.U:
-                _prevCenterLat = CenterLat;
-                _prevCenterLon = CenterLon;
-                _prevZoomLevel = ZoomLevel;
-
-                ZoomLevel = Math.Clamp(ZoomLevel / 1.25, 0.25, 20.0);
-                e.Handled = true;
-                ScheduleRadarRefresh();
-                break;
-
-            case Key.W:
-                ShowWeather = !ShowWeather;
-                if (ShowWeather)
-                    ScheduleRadarRefresh();
-                e.Handled = true;
-                break;
-
-            case Key.X:
-                var tempLat = CenterLat;
-                var tempLon = CenterLon;
-                var tempZoom = ZoomLevel;
-
-                CenterLat = _prevCenterLat;
-                CenterLon = _prevCenterLon;
-                ZoomLevel = _prevZoomLevel;
-
-                _prevCenterLat = tempLat;
-                _prevCenterLon = tempLon;
-                _prevZoomLevel = tempZoom;
-
-                e.Handled = true;
-                ScheduleRadarRefresh();
-                break;
-
-            case Key.Z:
-                _prevCenterLat = CenterLat;
-                _prevCenterLon = CenterLon;
-                _prevZoomLevel = ZoomLevel;
-
-                ZoomLevel = Math.Clamp(ZoomLevel * 1.25, 0.25, 20.0);
-                e.Handled = true;
-                ScheduleRadarRefresh();
-                break;
-        }
     }
 
     // =========================================================================
@@ -1216,648 +353,5 @@ public class TsdRadarControl : Control
         }
 
         return geo;
-    }
-
-    // =========================================================================
-    // Render — main entry point
-    // =========================================================================
-
-    public override void Render(DrawingContext context)
-    {
-        var width = Bounds.Width;
-        var height = Bounds.Height;
-        if (width <= 0 || height <= 0) return;
-
-        // Background
-        context.FillRectangle(
-            _backgroundBrush, new Rect(0, 0, width, height));
-
-        // Rebuild cached geometries if stale
-        if (_geometriesDirty ||
-            Math.Abs(_cachedWidth - width) > 0.5 ||
-            Math.Abs(_cachedHeight - height) > 0.5)
-        {
-            RebuildGeometries(width, height);
-        }
-
-        // Boundaries
-        if (ShowStateBoundaries && _stateBoundaryGeometry != null)
-            context.DrawGeometry(
-                null, _boundaryPen, _stateBoundaryGeometry);
-
-        if (ShowCountryBoundaries && _countryBoundaryGeometry != null)
-            context.DrawGeometry(
-                null, _boundaryPen, _countryBoundaryGeometry);
-
-        if (_sectorGeometry != null)
-            context.DrawGeometry(null, _sectorPen, _sectorGeometry);
-
-        // Weather radar overlay
-        if (ShowWeather && _radarBitmap != null)
-        {
-            var topLeft = LatLonToScreen(
-                RadarMaxLat, RadarMinLon, width, height);
-            var bottomRight = LatLonToScreen(
-                RadarMinLat, RadarMaxLon, width, height);
-
-            var destRect = new Rect(
-                topLeft.X, topLeft.Y,
-                bottomRight.X - topLeft.X,
-                bottomRight.Y - topLeft.Y);
-
-            using (context.PushOpacity(RadarOpacity))
-            {
-                context.DrawImage(_radarBitmap,
-                    new Rect(0, 0,
-                        _radarBitmap.PixelSize.Width,
-                        _radarBitmap.PixelSize.Height),
-                    destRect);
-            }
-        }
-
-        // ARTCC boundaries
-        if (ShowArtcc && _artccBoundaryGeometry != null)
-            context.DrawGeometry(null, _artccPen, _artccBoundaryGeometry);
-
-        // Overlays and flights
-        DrawRangeRings(context, width, height);
-        DrawRoutes(context, width, height);
-        DrawActiveMapItems(context, width, height);
-        DrawAircraft(context, width, height);
-    }
-
-    // =========================================================================
-    // Draw — Aircraft and data blocks
-    // =========================================================================
-
-    private void DrawAircraft(DrawingContext context,
-                               double width, double height)
-    {
-        var typeface = _dataBlockTypeface;
-        const double size = 4.0;
-
-        foreach (var pilot in VisiblePilots)
-        {
-            if (pilot.IsHidden) continue;
-
-            var pt = LatLonToScreen(
-                pilot.Lat, pilot.Lon, width, height);
-            if (!IsOnScreen(pt, width, height)) continue;
-
-            var color = pilot.ColorOverride ?? pilot.MatchedFilterColor;
-            var brush = new SolidColorBrush(Color.Parse(color));
-
-            DrawAircraftSymbol(context, pt, pilot.Heading, size, brush);
-
-            // Show data block if persistent, or on hover
-            if (pilot.ShowDataBlock ||
-                (_hoveredCallsign != null &&
-                 _hoveredCallsign == pilot.Callsign))
-            {
-                DrawDataBlock(context, pt, pilot,
-                    _dataBlockBrush, typeface);
-            }
-        }
-    }
-
-    private void DrawAircraftSymbol(DrawingContext context,
-        Point pt, int heading, double size, IBrush brush)
-    {
-        double scale = size / 6.0;
-        double rad = heading * Math.PI / 180.0;
-
-        Point Transform(double x, double y)
-        {
-            double rx = x * Math.Cos(rad) - y * Math.Sin(rad);
-            double ry = x * Math.Sin(rad) + y * Math.Cos(rad);
-            return new Point(pt.X + rx * scale, pt.Y + ry * scale);
-        }
-
-        var geo = new StreamGeometry();
-        using (var ctx = geo.Open())
-        {
-            ctx.BeginFigure(Transform(0, -6.5), true);
-            ctx.LineTo(Transform(1, -8));
-            ctx.LineTo(Transform(1, -2));
-            ctx.LineTo(Transform(8, 4));
-            ctx.LineTo(Transform(6, 5));
-            ctx.LineTo(Transform(1, 1));
-            ctx.LineTo(Transform(1, 6));
-            ctx.LineTo(Transform(4, 9));
-            ctx.LineTo(Transform(4, 10));
-            ctx.LineTo(Transform(1, 8));
-            ctx.LineTo(Transform(0, 8));
-            ctx.LineTo(Transform(-1, 8));
-            ctx.LineTo(Transform(-4, 10));
-            ctx.LineTo(Transform(-4, 9));
-            ctx.LineTo(Transform(-1, 6));
-            ctx.LineTo(Transform(-1, 1));
-            ctx.LineTo(Transform(-6, 5));
-            ctx.LineTo(Transform(-8, 4));
-            ctx.LineTo(Transform(-1, -2));
-            ctx.LineTo(Transform(-1, -8));
-            ctx.EndFigure(true);
-        }
-
-        context.DrawGeometry(brush, null, geo);
-    }
-
-    private void DrawDataBlock(DrawingContext context,
-        Point pt, VatsimPilot pilot,
-        IBrush textBrush, Typeface typeface)
-    {
-        string altStr = pilot.Altitude >= 18000
-            ? $"F{pilot.Altitude / 100:000}"
-            : $"{pilot.Altitude / 100:000}";
-
-        var linesList = new List<string>
-        {
-            pilot.Callsign,
-            $"{pilot.AircraftType,-4} {altStr}",
-            $"{pilot.GroundSpeed}"
-        };
-
-        if (pilot.ShowOrgDest)
-            linesList.Add($"{pilot.Departure} → {pilot.Arrival}");
-        else
-            linesList.Add(pilot.Arrival);
-
-        if ((pilot.MatchedShowRoute || pilot.ManualShowRoute) &&
-            !string.IsNullOrWhiteSpace(pilot.Route))
-        {
-            const int wrapWidth = 30;
-            var route = pilot.Route;
-            while (route.Length > 0)
-            {
-                if (route.Length <= wrapWidth)
-                {
-                    linesList.Add(route);
-                    break;
-                }
-
-                int breakAt = route.LastIndexOf(' ', wrapWidth);
-                if (breakAt <= 0) breakAt = wrapWidth;
-
-                linesList.Add(route[..breakAt].Trim());
-                route = route[breakAt..].Trim();
-            }
-        }
-
-        var lines = linesList.ToArray();
-        double lineHeight = 12.0;
-        double bx = pt.X + 10;
-        double by = pt.Y - (lines.Length * lineHeight) / 2;
-
-        for (int i = 0; i < lines.Length; i++)
-        {
-            var text = new FormattedText(
-                lines[i],
-                System.Globalization.CultureInfo.CurrentCulture,
-                FlowDirection.LeftToRight,
-                typeface, 10, textBrush);
-
-            context.DrawText(text, new Point(bx, by + i * lineHeight));
-        }
-    }
-
-    // =========================================================================
-    // Draw — Flight routes
-    // =========================================================================
-
-    private void DrawRoutes(DrawingContext context,
-        double width, double height)
-    {
-        foreach (var pilot in VisiblePilots)
-        {
-            if (!pilot.MatchedDrawRoute && !pilot.ManualDrawRoute) continue;
-            if (pilot.IsHidden) continue;
-            if (pilot.ParsedRoute is null ||
-                pilot.ParsedRoute.Count < 2) continue;
-
-            var brush = new SolidColorBrush(
-                Color.Parse(pilot.ColorOverride ?? pilot.MatchedFilterColor));
-            var pen = new Pen(brush, 1.0);
-
-            int nextWaypointIndex = FindNextWaypoint(
-                pilot.Lat, pilot.Lon, pilot.ParsedRoute);
-
-            var geo = new StreamGeometry();
-            using (var ctx = geo.Open())
-            {
-                var aircraftPt = LatLonToScreen(
-                    pilot.Lat, pilot.Lon, width, height);
-                ctx.BeginFigure(aircraftPt, false);
-
-                for (int i = nextWaypointIndex;
-                    i < pilot.ParsedRoute.Count; i++)
-                {
-                    if (pilot.ParsedRoute[i] is null) continue;
-                    ctx.LineTo(LatLonToScreen(
-                        pilot.ParsedRoute[i]!.Lat,
-                        pilot.ParsedRoute[i]!.Lon,
-                        width, height));
-                }
-                ctx.EndFigure(false);
-            }
-
-            context.DrawGeometry(null, pen, geo);
-        }
-    }
-
-    private static int FindNextWaypoint(
-        double pilotLat, double pilotLon,
-        List<LatLon> route)
-    {
-        double minDist = double.MaxValue;
-        int closestIndex = 0;
-
-        for (int i = 0; i < route.Count; i++)
-        {
-            double dLat = route[i].Lat - pilotLat;
-            double dLon = route[i].Lon - pilotLon;
-            double dist = dLat * dLat + dLon * dLon;
-
-            if (dist < minDist)
-            {
-                minDist = dist;
-                closestIndex = i;
-            }
-        }
-
-        return Math.Min(closestIndex + 1, route.Count - 1);
-    }
-
-    // =========================================================================
-    // Draw — Range rings
-    // =========================================================================
-
-    private void DrawRangeRings(DrawingContext context,
-        double width, double height)
-    {
-        if (RangeRings.Count == 0) return;
-
-        var pen = new Pen(new SolidColorBrush(
-            Color.Parse("#888888")), 0.8);
-        var labelBrush = new SolidColorBrush(
-            Color.Parse("#888888"));
-
-        foreach (var config in RangeRings)
-        {
-            int ringCount = config.DistanceNm / config.IntervalNm;
-
-            for (int r = 1; r <= ringCount; r++)
-            {
-                double radiusNm = r * config.IntervalNm;
-                double radiusDegLat = radiusNm / 60.0;
-                double radiusDegLon = radiusDegLat /
-                    Math.Cos(config.CenterLat * Math.PI / 180.0);
-
-                var geo = new StreamGeometry();
-                using (var ctx = geo.Open())
-                {
-                    const int segments = 72;
-                    for (int i = 0; i <= segments; i++)
-                    {
-                        double angle = 2.0 * Math.PI * i / segments;
-                        double lat = config.CenterLat +
-                            radiusDegLat * Math.Sin(angle);
-                        double lon = config.CenterLon +
-                            radiusDegLon * Math.Cos(angle);
-
-                        var pt = LatLonToScreen(lat, lon, width, height);
-
-                        if (i == 0)
-                            ctx.BeginFigure(pt, false);
-                        else
-                            ctx.LineTo(pt);
-                    }
-                    ctx.EndFigure(true);
-                }
-
-                context.DrawGeometry(null, pen, geo);
-
-                // Distance label at top of ring
-                double labelLat = config.CenterLat + radiusDegLat;
-                var labelPt = LatLonToScreen(
-                    labelLat, config.CenterLon, width, height);
-
-                var ringLabel = new FormattedText(
-                    $"{radiusNm}nm",
-                    System.Globalization.CultureInfo.CurrentCulture,
-                    FlowDirection.LeftToRight,
-                    _mapLabelTypeface, 12, labelBrush);
-
-                context.DrawText(ringLabel, new Point(
-                    labelPt.X - ringLabel.Width / 2,
-                    labelPt.Y - ringLabel.Height - 2));
-            }
-
-            // Center dot and identifier
-            var centerPt = LatLonToScreen(
-                config.CenterLat, config.CenterLon, width, height);
-
-            context.FillRectangle(labelBrush,
-                new Rect(centerPt.X - 2, centerPt.Y - 2, 4, 4));
-
-            var label = new FormattedText(
-                config.Identifier,
-                System.Globalization.CultureInfo.CurrentCulture,
-                FlowDirection.LeftToRight,
-                _mapLabelTypeface, 9, labelBrush);
-
-            context.DrawText(label, new Point(
-                centerPt.X + 4, centerPt.Y - label.Height / 2));
-        }
-    }
-
-    // =========================================================================
-    // Draw — Active map items (airports, VORs, NDBs, fixes, TRACONs, etc.)
-    // =========================================================================
-
-    private void DrawActiveMapItems(DrawingContext context,
-                                     double width, double height)
-    {
-        const double half = 2.5;
-        const double triSize = 5.0;
-
-        foreach (var item in ActiveMapItems)
-        {
-            var pt = LatLonToScreen(item.Lat, item.Lon, width, height);
-            if (item.Type != "Airway" && item.Type != "Sector" &&
-                !IsOnScreen(pt, width, height)) continue;
-
-            switch (item.Type)
-            {
-                case "Airport":
-                    context.FillRectangle(_airportBrush,
-                        new Rect(pt.X - half, pt.Y - half,
-                            half * 2, half * 2));
-
-                    if (item.ShowLabel)
-                    {
-                        var airportLabel = new FormattedText(
-                            item.Identifier.Length > 3
-                                ? item.Identifier[..3]
-                                : item.Identifier,
-                            System.Globalization.CultureInfo.CurrentCulture,
-                            FlowDirection.LeftToRight,
-                            _mapLabelTypeface, 9, _airportBrush);
-
-                        context.DrawText(airportLabel,
-                            new Point(pt.X + half + 2,
-                                pt.Y - airportLabel.Height / 2));
-                    }
-                    break;
-
-                case string t when t.Contains("VOR"):
-                    context.DrawEllipse(null, _vorPen, pt, 2.5, 4.0);
-
-                    if (item.ShowLabel)
-                    {
-                        var vorLabel = new FormattedText(
-                            item.Identifier.Length > 3
-                                ? item.Identifier[..3]
-                                : item.Identifier,
-                            System.Globalization.CultureInfo.CurrentCulture,
-                            FlowDirection.LeftToRight,
-                            _mapLabelTypeface, 9, _vorBrush);
-
-                        context.DrawText(vorLabel,
-                            new Point(pt.X + 6,
-                                pt.Y - vorLabel.Height / 2));
-                    }
-                    break;
-
-                case string t when t.Contains("NDB"):
-                    context.DrawEllipse(null, _ndbPen, pt, 2.5, 4.0);
-
-                    if (item.ShowLabel)
-                    {
-                        var ndbLabel = new FormattedText(
-                            item.Identifier.Length > 3
-                                ? item.Identifier[..3]
-                                : item.Identifier,
-                            System.Globalization.CultureInfo.CurrentCulture,
-                            FlowDirection.LeftToRight,
-                            _mapLabelTypeface, 9, _ndbBrush);
-
-                        context.DrawText(ndbLabel,
-                            new Point(pt.X + 6,
-                                pt.Y - ndbLabel.Height / 2));
-                    }
-                    break;
-
-                case "Fix":
-                    var top = new Point(pt.X, pt.Y - triSize);
-                    var bottomLeft = new Point(
-                        pt.X - triSize * 0.866, pt.Y + triSize * 0.5);
-                    var bottomRight = new Point(
-                        pt.X + triSize * 0.866, pt.Y + triSize * 0.5);
-
-                    var geo = new StreamGeometry();
-                    using (var ctx = geo.Open())
-                    {
-                        ctx.BeginFigure(top, false);
-                        ctx.LineTo(bottomLeft);
-                        ctx.LineTo(bottomRight);
-                        ctx.EndFigure(true);
-                    }
-                    context.DrawGeometry(null, _fixPen, geo);
-
-                    if (item.ShowLabel)
-                    {
-                        var fixLabel = new FormattedText(
-                            item.Identifier.Length > 5
-                                ? item.Identifier[..5]
-                                : item.Identifier,
-                            System.Globalization.CultureInfo.CurrentCulture,
-                            FlowDirection.LeftToRight,
-                            _mapLabelTypeface, 9, _fixBrush);
-
-                        context.DrawText(fixLabel,
-                            new Point(pt.X + triSize + 2,
-                                pt.Y - fixLabel.Height / 2));
-                    }
-                    break;
-
-                case "TRACON":
-                    DrawTraconBoundary(context, item, width, height);
-                    break;
-
-                case "Sector":
-                    // Draw sector boundary rings
-                    foreach (var ring in item.Rings)
-                    {
-                        if (ring.Count < 2) continue;
-
-                        var sectorGeo = new StreamGeometry();
-                        using (var ctx = sectorGeo.Open())
-                        {
-                            var first = LatLonToScreen(
-                                ring[0].Lat, ring[0].Lon, width, height);
-                            ctx.BeginFigure(first, false);
-                            for (int i = 1; i < ring.Count; i++)
-                            {
-                                ctx.LineTo(LatLonToScreen(
-                                    ring[i].Lat, ring[i].Lon, width, height));
-                            }
-                            ctx.EndFigure(true);
-                        }
-                        context.DrawGeometry(null, _sectorPen, sectorGeo);
-                    }
-
-                    // Sector labels (identifier + altitude)
-                    if (item.ShowLabel)
-                    {
-                        foreach (var labelRing in item.Rings)
-                        {
-                            if (labelRing.Count == 0) continue;
-
-                            double sLat = labelRing.Average(p => p.Lat);
-                            double sLon = labelRing.Average(p => p.Lon);
-                            var center = LatLonToScreen(
-                                sLat, sLon, width, height);
-
-                            if (!IsOnScreen(center, width, height)) continue;
-
-                            var idLabel = new FormattedText(
-                                item.Identifier,
-                                System.Globalization.CultureInfo.CurrentCulture,
-                                FlowDirection.LeftToRight,
-                                _mapLabelTypeface, 9, _mapLabelBrush);
-
-                            var altLabel = new FormattedText(
-                                item.Label,
-                                System.Globalization.CultureInfo.CurrentCulture,
-                                FlowDirection.LeftToRight,
-                                _mapLabelTypeface, 9, _mapLabelBrush);
-
-                            double totalHeight =
-                                idLabel.Height + altLabel.Height + 2;
-
-                            context.DrawText(idLabel, new Point(
-                                center.X - idLabel.Width / 2,
-                                center.Y - totalHeight / 2));
-
-                            context.DrawText(altLabel, new Point(
-                                center.X - altLabel.Width / 2,
-                                center.Y - totalHeight / 2 +
-                                    idLabel.Height + 2));
-                        }
-                    }
-                    break;
-
-                case "Airway":
-                    if (item.Rings.Count > 0)
-                    {
-                        // Draw airway line
-                        var airwayGeo = new StreamGeometry();
-                        using (var ctx = airwayGeo.Open())
-                        {
-                            var ring = item.Rings[0];
-                            if (ring.Count > 0)
-                            {
-                                var first = LatLonToScreen(
-                                    ring[0].Lat, ring[0].Lon,
-                                    width, height);
-                                ctx.BeginFigure(first, false);
-                                for (int i = 1; i < ring.Count; i++)
-                                {
-                                    ctx.LineTo(LatLonToScreen(
-                                        ring[i].Lat, ring[i].Lon,
-                                        width, height));
-                                }
-                                ctx.EndFigure(false);
-                            }
-                        }
-
-                        var airwayPen =
-                            item.Identifier.StartsWith("V",
-                                StringComparison.OrdinalIgnoreCase) ||
-                            item.Identifier.StartsWith("T",
-                                StringComparison.OrdinalIgnoreCase)
-                            ? _victorRoutePen : _jetRoutePen;
-
-                        context.DrawGeometry(null, airwayPen, airwayGeo);
-
-                        // Airway identifier label at midpoint
-                        if (item.ShowLabel)
-                        {
-                            var midPt =
-                                item.Rings[0][item.Rings[0].Count / 2];
-                            var midScreen = LatLonToScreen(
-                                midPt.Lat, midPt.Lon, width, height);
-
-                            if (IsOnScreen(midScreen, width, height))
-                            {
-                                var airwayLabel = new FormattedText(
-                                    item.Identifier,
-                                    System.Globalization.CultureInfo
-                                        .CurrentCulture,
-                                    FlowDirection.LeftToRight,
-                                    _mapLabelTypeface, 9, _mapLabelBrush);
-
-                                context.DrawText(airwayLabel, new Point(
-                                    midScreen.X - airwayLabel.Width / 2,
-                                    midScreen.Y - airwayLabel.Height / 2));
-                            }
-                        }
-                    }
-                    break;
-            }
-        }
-    }
-
-    // =========================================================================
-    // Draw — TRACON boundary (special case with multi-ring + centered label)
-    // =========================================================================
-
-    private void DrawTraconBoundary(DrawingContext context,
-        MapItem item, double width, double height)
-    {
-        foreach (var ring in item.Rings)
-        {
-            if (ring.Count < 2) continue;
-
-            var geometry = new StreamGeometry();
-            using (var ctx = geometry.Open())
-            {
-                var first = LatLonToScreen(
-                    ring[0].Lat, ring[0].Lon, width, height);
-                ctx.BeginFigure(first, false);
-
-                for (int i = 1; i < ring.Count; i++)
-                {
-                    var pt = LatLonToScreen(
-                        ring[i].Lat, ring[i].Lon, width, height);
-                    ctx.LineTo(pt);
-                }
-                ctx.EndFigure(true);
-            }
-
-            context.DrawGeometry(null, _traconPen, geometry);
-        }
-
-        if (item.ShowLabel)
-        {
-            var allPoints = item.Rings.SelectMany(r => r).ToList();
-            if (allPoints.Count > 0)
-            {
-                double avgLat = allPoints.Average(p => p.Lat);
-                double avgLon = allPoints.Average(p => p.Lon);
-                var center = LatLonToScreen(
-                    avgLat, avgLon, width, height);
-
-                var label = new FormattedText(
-                    item.Identifier,
-                    System.Globalization.CultureInfo.CurrentCulture,
-                    FlowDirection.LeftToRight,
-                    _mapLabelTypeface, 9, _traconBrush);
-
-                context.DrawText(label, new Point(
-                    center.X - label.Width / 2,
-                    center.Y - label.Height / 2));
-            }
-        }
     }
 }
